Isolated web socket handlers
